Created KMeansCommons.KMeansCommons trait for common functions in private KMeans classes. Moved KMeansMiniBatch.{initRandom, initKMeansMiniBatchParallel} there
