+++ mllib/src/main/scala/org/apache/spark/mllib/clustering/KMeansMiniBatch.scala
@@ -44,7 +44,7 @@ class KMeansMiniBatch private (
     private var runs: Int,
     private var initializationMode: String,
     private var initializationSteps: Int,
-    private var epsilon: Double) extends Serializable with Logging {
+    private var epsilon: Double) extends Serializable with KMeansCommons with Logging {
 
   /**
    * Constructs a KMeans instance with default parameters: {k: 2, maxIterations: 20, runs: 1,
@@ -150,9 +150,9 @@ class KMeansMiniBatch private (
     val initStartTime = System.nanoTime()
 
     val centers = if (initializationMode == KMeansMiniBatch.RANDOM) {
-      initRandom(data)
+      initRandom(data, k)
     } else {
-      initKMeansMiniBatchParallel(data)
+      initKMeansMiniBatchParallel(data, k, initializationSteps)
     }
 
     val initTimeInSeconds = (System.nanoTime() - initStartTime) / 1e9
@@ -215,69 +215,6 @@ class KMeansMiniBatch private (
     logInfo(s"The cost for the run is $costs.")
 
     new Tuple2(new KMeansModel(centers.map(c => Vectors.fromBreeze(c.vector))), costs)
-  }
-
-  /**
-   * Initialize `runs` sets of cluster centers at random.
-   */
-  private def initRandom(data: RDD[BreezeVectorWithNorm])
-  : Array[BreezeVectorWithNorm] = {
-    // Sample all the cluster centers in one pass to avoid repeated scans
-    val sample = data.takeSample(true, k, new XORShiftRandom().nextInt()).toSeq
-    sample.map { v =>
-      new BreezeVectorWithNorm(v.vector.toDenseVector, v.norm)
-    }.toArray
-  }
-
-  /**
-   * Initialize `runs` sets of cluster centers using the k-means|| algorithm by Bahmani et al.
-   * (Bahmani et al., Scalable K-Means++, VLDB 2012). This is a variant of k-means++ that tries
-   * to find with dissimilar cluster centers by starting with a random center and then doing
-   * passes where more centers are chosen with probability proportional to their squared distance
-   * to the current cluster set. It results in a provable approximation to an optimal clustering.
-   *
-   * The original paper can be found at http://theory.stanford.edu/~sergei/papers/vldb12-kmpar.pdf.
-   */
-  private def initKMeansMiniBatchParallel(data: RDD[BreezeVectorWithNorm])
-  : Array[BreezeVectorWithNorm] = {
-    // Initialize each run's center to a random point
-    val seed = new XORShiftRandom().nextInt()
-    val sample = data.takeSample(true, 1, seed).toSeq
-    val centers = ArrayBuffer() ++ sample
-
-    // On each step, sample 2 * k points on average for each run with probability proportional
-    // to their squared distance from that run's current centers
-    var step = 0
-    while (step < initializationSteps) {
-      val sumCosts = data.map { point =>
-      		KMeansMiniBatch.pointCost(centers, point)
-      }.reduce(_ + _)
-      val chosen = data.mapPartitionsWithIndex { (index, points) =>
-        val rand = new XORShiftRandom(seed ^ (step << 16) ^ index)
-        
-        // accept / reject each point
-        val sampledCenters = points.filter { p =>
-          rand.nextDouble() < 2.0 * KMeansMiniBatch.pointCost(centers, p) * k / sumCosts
-        }
-        
-        sampledCenters
-      }.collect()
-      
-      
-      centers ++= chosen
-      step += 1
-    }
-    
-    // Finally, we might have a set of more than k candidate centers for each run; weigh each
-    // candidate by the number of points in the dataset mapping to it and run a local k-means++
-    // on the weighted centers to pick just k of them
-    val weightMap = data.map { p =>
-        (KMeansMiniBatch.findClosest(centers, p)._1, 1.0)
-      }.reduceByKey(_ + _).collectAsMap()
-    val weights = (0 until centers.length).map(i => weightMap.getOrElse(i, 0.0)).toArray
-    val finalCenters = LocalKMeans.kMeansPlusPlus(seed, centers.toArray, weights, k, 30)
-
-    finalCenters.toArray
   }
 }
